Move SimplisticHistory to HistoryInstances (RCHAIN-3415)
